// packages/job/src/knowledge/wikimatch/types.ts
export interface WikimatchCandidate {
  readonly hasMoreOptions?: boolean;
  readonly id: string;
  readonly qidOptions: readonly WikimatchQidOption[];
  readonly range: WikimatchTextRange;
  readonly surface: string;
}

export interface WikimatchSentence {
  readonly id?: string;
  readonly range: WikimatchTextRange;
  readonly text: string;
}

export interface WikimatchQidOption {
  readonly description?: string;
  readonly disambiguation?: readonly WikimediaDisambiguationItem[];
  readonly isDisambiguation?: boolean;
  readonly label?: string;
  readonly qid: string;
  readonly url?: string;
}

export interface WikimatchTextRange {
  readonly end: number;
  readonly start: number;
}